requirements.txt
qdrant-client>=1.7.0
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
plotly>=5.17.0
# optional: openai>=1.0.0  (only if EMBEDDING_MODEL=openai)
